# target/classes/specifications/mcs-card-api-swagger.yaml
openapi: 3.0.0
info:
  description: |
    API definitions for Mastercard Secure Card on File Cards API Endpoint. The Cards endpoint enrolls a mastercard card into a specified Secure Card on File. The Get Card by ID will retrieve the MaskedCard details, for the specified card ID. Delete card by card ID will delete the previously enrolled card.
  version: 1.0.0
  title: Mastercard Checkout Solutions Card API
  termsOfService: https://developer.mastercard.com/terms-of-use
  contact:
    email: apisupport@mastercard.com
    name: API Support
    url: https://developer.mastercard.com/support
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
...
servers:
  - description: Production Server
    url: https://api.mastercard.com/src/api/digital/payments
  - description: Sandbox server
    url: https://sandbox.api.mastercard.com/src/api/digital/payments
paths:
...
  /cards/{cardId}:
    get:
      summary: Get card by card ID
      description: |-
        The GetCardRequest is used to retrieve the latest card information. It must contain the following values: <ul> <li>serviceId </li> <li>srcClientId </li> <li>srcDigitalCardId </li> </ul>
        The following values are conditional: <ul> <li>srcDpaId </li> </ul>
        <br>The parameters passed in the HTTP request header must be in an encoded format to be valid for HTTP transmission.
      operationId: getCard
      parameters:
        - in: path
          name: cardId
          description: The srcDigitalCardId of the card enrolled in Mastercard Checkout Solutions (MCS).
          example: "9ajNArkAR66zqYyHCn0rCA000000000000US"
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/SrcClientId'
        - $ref: '#/components/parameters/SrcDpaId'
        - $ref: '#/components/parameters/SrcCorrelationId'
        - $ref: '#/components/parameters/SrciTransactionId'
        - $ref: '#/components/parameters/ServiceId'
        - in: query
          name: tokenRequested
          description: |
            This value is defaulted to "false". The ability to receive an encrypted payment token object in the get card response is manually controlled by Mastercard. Integrators must work with their account managers to be able to receive a token in the get card response. Once permitted and enabled, set this value to "true".
          example: false
          schema:
            type: boolean
            default: false
        - in: query
          name: keyFingerprintId
          description: Mastercard will use the encryption key associated with the supplied KID to encrypt the payload. If not supplied, the key which is marked as default will be used.
          example: "a28f0222ee06631fefdea53b61ece178a593725142df927889ddcd95680baab4"
          schema:
            type: string
        - in: header
          name: X-Src-Cx-Flow-Id
          description: "The X-Src-Cx-Flow-Id ensures that back-to-back API calls will be sent to the same Mastercard server to maintain session affinity. When you make your initial API call, Secure Card on File responds with the X-Src-Cx-Flow-Id in the header. The X-Src-Cx-Flow-Id returned in this response should be used in the next back-to-back call to ensure these calls are directed to the same server. If the same X-Src-Cx-Flow-Id is not used for back-to-back calls, you may encounter site replication errors."
          schema:
            maxLength: 255
            pattern: "^[0-9a-zA-Z-_#.]*$"
            type: string
            example: "34F4A04B.B9EF8486-E8C1-44F3-88DC-D4F516A354DC"
      responses:
        "200":
          description: |
            OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetCardResponse'
          headers:
            X-Src-Cx-Flow-Id:
              description: "The X-Src-Cx-Flow-Id ensures that back-to-back API calls will be sent to the same Mastercard server to maintain session affinity. When you make your initial API call, Secure Card on File responds with the X-Src-Cx-Flow-Id in the header. The X-Src-Cx-Flow-Id returned in this response should be used in the next back-to-back call to ensure these calls are directed to the same server. If the same X-Src-Cx-Flow-Id is not used for back-to-back calls, you may encounter site replication errors."
              example: "4f339be7.17edcbd5-213a-4388-94ad-8c508d5f71ae.1682972398"
              schema:
                maxLength: 255
                pattern: "^[0-9a-zA-Z-_#]*$"
                type: string
        "400":
          content:
            application/json:
              example:
                errordetail:
                  - message: "Failed to convert value of type 'java.lang.String' to required type 'java.lang.Boolean'; nested exception is java.lang.IllegalArgumentException: Invalid boolean value [falsetrue]"
                    reason: INVALID_VALUE
                    source: tokenRequested
                    sourceType: QUERY
                message: Http request argument type is mismatched at this service
                reason: INVALID_ARGUMENT
                status: 400
              schema:
                $ref: '#/components/schemas/Error'
          description: |
            Bad Request. Invalid tokenRequested field.
        "403":
          content:
            application/json:
              examples:
                PERMISSION_DENIED:
                  description: |
                    'Srci not on-boarded/ServiceId is invalid/DpaId is not onboarded/ Requested dpaId is not associated with Srci.'
                  value:
                    errordetail: null
                    message: Access is denied to the requested resource. Client does not have sufficient permission or has been locked.
                    reason: PERMISSION_DENIED
                    status: 403
                PERMISSION_DENIED-EXPIRED:
                  description: |
                    See the message content for extra information. Srci returned expired encryption certificate.
                  value:
                    errordetail:
                      - message: The Public Encryption Key for this Client has expired. Please contact Mastercard administrator.
                        reason: EXPIRED
                        source: srcClientId
                        sourceType: BODY
                    message: Access is denied to the requested resource. Client does not have sufficient permission or has been locked.
                    reason: PERMISSION_DENIED
                    status: 403
                PERMISSION_DENIED-INVALID_VALUE:
                  description: |
                    See the message content for extra information. For example, requested program does not have authorization to access token data.
                  value:
                    errordetail:
                      - message: Program represented by Service-id is not enabled for serving Token Data. Please contact Mastercard administrator to enable this feature.
                        reason: INVALID_VALUE
                        source: serviceId
                        sourceType: BODY
                    message: Access is denied to the requested resource. Client does not have sufficient permission or has been locked.
                    reason: PERMISSION_DENIED
                    status: 403
                PERMISSION_DENIED-NOT_FOUND:
                  description: |
                    no active key returned from tcs / in_use key invalid / fringerPrints not matched
                  value:
                    errordetail:
                      - message: No active key found for this Client.
                        reason: NOT_FOUND
                        source: srcClientId
                        sourceType: BODY
                    message: Access is denied to the requested resource. Client does not have sufficient permission or has been locked.
                    reason: PERMISSION_DENIED
                    status: 403
              schema:
                $ref: '#/components/schemas/Error'
          description: |
            Forbidden.
        "404":
          content:
            application/json:
              example:
                errordetail:
                  - message: Card ID not found
                    reason:  NOT_FOUND
                    source: srcDigitalCardId
                    sourceType: PATH
                message: Not Found
                reason: NOT_FOUND
                status: 404
              schema:
                $ref: '#/components/schemas/Error'
          description: |
            Card not found by cards service.
        "500":
          $ref: '#/components/responses/InternalServerError'
        "503":
          $ref: '#/components/responses/ServiceUnavailable'
      tags:
        - Cards
    delete:
      summary: Delete card by card ID.
      description: Delete card by card ID The Delete Card by ID operation will delete the card with the specified card identifier. Delete a previously enrolled card with mastercard.<br>The parameters passed in the HTTP request header must be in an encoded format to be valid for HTTP transmission.
      operationId: deleteCard
      parameters:
        - in: path
          name: cardId
          description: The srcDigitalCardId of the card enrolled in Mastercard Checkout Solutions (MCS).
          example: "n8TdCz6BQvSIj2C0yYPvZw000000000000US"
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/SrcClientId'
        - $ref: '#/components/parameters/SrcDpaId'
        - $ref: '#/components/parameters/SrcCorrelationId'
        - $ref: '#/components/parameters/SrciTransactionId'
        - $ref: '#/components/parameters/ServiceId'
        - in: query
          name: reason
          description: Reason for the card deletion.
          example: "ACCOUNT_CLOSED"
          schema:
            type: string
            enum:
              - SUSPECTED_FRAUD
              - ACCOUNT_CLOSED
              - OTHER
              - DEFAULT
        - in: query
          name: requestor
          description: Entity initiating the card deletion.
          example: "CARDHOLDER"
          schema:
            type: string
            enum:
              - CARDHOLDER
              - MERCHANT
              - ISSUER
              - WALLET
      responses:
        "200":
          description: |
            OK, Returned when the card and its dependencies has been successfully deleted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteCardResponse'
          headers:
            X-Src-Cx-Flow-Id:
              description: "The X-Src-Cx-Flow-Id ensures that back-to-back API calls will be sent to the same Mastercard server to maintain session affinity. When you make your initial API call, Secure Card on File responds with the X-Src-Cx-Flow-Id in the header. The X-Src-Cx-Flow-Id returned in this response should be used in the next back-to-back call to ensure these calls are directed to the same server. If the same X-Src-Cx-Flow-Id is not used for back-to-back calls, you may encounter site replication errors."
              schema:
                maxLength: 255
                pattern: "^[0-9a-zA-Z-_#.]*$"
                type: string
                example: "34F4A04B.B9EF8486-E8C1-44F3-88DC-D4F516A354DC"
            X-Src-Response-Host:
              description: "The X-Src-Response-Host is used to direct all calls from the same client to the same server to maintain session affinity. When you make your initial API call, Mastercard Checkout Solutions responds with the X-Src-Response-Host in the HTTP header. The X-Src-Response-Host that is returned will contain a site Specific URL that must be used in the subsequent calls to ensure these calls are directed to the same server that returned the initial response."
              schema:
                maxLength: 1024
                type: string
                example: "https://ksc.services-asn.mastercard.com/"
        "400":
          description: |
            Bad Request. Reason code of `INVALID_ARGUMENT` Identifies a malformed or invalid request.
            Use cases:
              * Mandatory Query parameters are missing example 'srcClientId' is not provided.
              * Query parameters provided value has invalid format example, 'srcCorrelationId' is not in UUID format.

              Reason code of `INVALID_STATE` identifies a validation failed on request. Use cases:
              * `srcClientId` or `srcDpaId` or `serviceId` is invalid and does not resolve to an entity in SRC system. Example `srcClientId` is not recognised.
              * Request Body JSON schema validation failure.
              * Data Type validation failure.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                INVALID_ARGUMENT-CANNOT_BE_NULL:
                  description: |
                    See the message content for extra information. Possible reasons are:

                      * digitalAccountReferenceValue as null
                      * null clientId'
                  value:
                    errordetail:
                      - message: Required request parameter 'srcClientId' for method parameter type String is not present
                        reason: CANNOT_BE_NULL
                        source: srcClientId
                        sourceType: QUERY
                    message: Http request query parameter is missing
                    reason: INVALID_ARGUMENT
                    status: 400
                INVALID_ARGUMENT-INVALID_VALUE:
                  description: |
                    'See the message content for extra information. Possible reasons are:

                      * invalid correlationId format
                      * invalid digitalAccountReferenceType
                      * invalid clientId'
                  value:
                    errordetail:
                      - message: "must match \"[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}\""
                        reason: INVALID_VALUE
                        source: srcClientId
                        sourceType: QUERY
                    message: Cannot process the request because it is malformed or has incorrect/missing fields or values.
                    reason: INVALID_ARGUMENT
                    status: 400
        "403":
          content:
            application/json:
              examples:
                PERMISSION_DENIED:
                  description: |
                    'See the message content for extra information. Possible reasons are invalid dpaId / invalid srci / invalid service id.'
                  value:
                    errordetail: null
                    message: Access is denied to the requested resource. Client does not have sufficient permission or has been locked.
                    reason: PERMISSION_DENIED
                    status: 403
                PERMISSION_DENIED-INVALID_VALUE:
                  description: |
                    See the message content for extra information. Possible reasons are:

                      * Access is denied to the requested resource. Client does not have sufficient permission or has been locked.
                      * invalid X_OPEN_API_CLIENT_ID'
                  value:
                    errordetail:
                      - message: Invalid X_OPEN_API_CLIENT_ID
                        reason: INVALID_VALUE
                        source: X_OPEN_API_CLIENT_ID
                        sourceType: HEADER
                    message: Access is denied to the requested resource. Client does not have sufficient permission or has been locked.
                    reason: PERMISSION_DENIED
                    status: 403
              schema:
                $ref: '#/components/schemas/Error'
          description: |
            Forbidden. API will return this response if identifier validation fails.
        "404":
          content:
            application/json:
              example:
                errordetail: null
                message: Card not found
                reason: NOT_FOUND
                status: 404
              schema:
                $ref: '#/components/schemas/Error'
          description: |
            Not Found. Identifies a missing Path Parameter. Possible reasons:

              * `cardId` is not provided.
              * `cardId` does not exist in the system.
        "500":
          $ref: '#/components/responses/InternalServerError'
        "503":
          $ref: '#/components/responses/ServiceUnavailable'
      tags:
        - Cards
  /cards/{cardId}/transactions:
    get:
      summary: RETRIEVE TRANSACTIONS
      description: "Get transactions by Card Id API will retrieve the recent transactions for the specified card identifier. Integrators may retrieve recent transactions details for a specified card identifier via the Get Transaction Details request. If the Issuer of the card does not support Transaction Detail Service, the Get Transaction Details request will return an error."
      operationId: getTransactionHistory
      parameters:
        - in: path
          name: cardId
          description: Unique identifier of the card. The card identifier may be an SRC Digital Card Identifier or an SRC Payment Card Identifier
          example: "9ajNArkAR66zqYyHCn0rCA000000000000US"
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/SrcClientId'
        - $ref: '#/components/parameters/ServiceId'
        - $ref: '#/components/parameters/SrciTransactionId'
        - in: query
          name: transactionsFromTimestamp
          description: Optional Parameter to provide Transaction history records originating after Date-Time represented by this parameter. This parameter MUST not past 30 days from current date-time and also not after current date-time. Supports taking value in UTC epoch milliseconds format.
          example: "2016-05-04T12:08:56.123-07:00"
          schema:
            type: string
            pattern: "^(-?(?:[1-9][0-9]*)?[0-9]{4})-(1[0-2]|0[1-9])-(3[01]|0[1-9]|[12][0-9])T(2[0-3]|[01][0-9]):([0-9][0-9]):([0-9][0-9]).([0-9][0-9][0-9])([-+])([0-9][0-9]):([0-9][0-9])?$"
      responses:
        "200":
          description: |
            OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetRecentTransactionsResponse'
        "400":
          content:
            application/json:
              example:
                errordetail:
                  - message: "Failed to convert value of type 'java.lang.String' to required type 'java.lang.Boolean'; nested exception is java.lang.IllegalArgumentException: Invalid boolean value [falsetrue]"
                    reason: INVALID_VALUE
                    source: tokenRequested
                    sourceType: QUERY
                message: Http request argument type is mismatched at this service
                reason: INVALID_ARGUMENT
                status: 400
              schema:
                $ref: '#/components/schemas/Error'
          description: |
            Bad Request. Invalid tokenRequested field.
        "403":
          content:
            application/json:
              examples:
                PERMISSION_DENIED:
                  description: |
                    'Srci not on-boarded/ServiceId is invalid/DpaId is not onboarded/ Requested dpaId is not associated with Srci.'
                  value:
                    errordetail: null
                    message: Access is denied to the requested resource. Client does not have sufficient permission or has been locked.
                    reason: PERMISSION_DENIED
                    status: 403
                PERMISSION_DENIED-EXPIRED:
                  description: |
                    See the message content for extra information. Srci returned expired encryption certificate.
                  value:
                    errordetail:
                      - message: The Public Encryption Key for this Client has expired. Please contact Mastercard administrator.
                        reason: EXPIRED
                        source: srcClientId
                        sourceType: BODY
                    message: Access is denied to the requested resource. Client does not have sufficient permission or has been locked.
                    reason: PERMISSION_DENIED
                    status: 403
                PERMISSION_DENIED-INVALID_VALUE:
                  description: |
                    See the message content for extra information. For example, requested program does not have authorization to access token data.
                  value:
                    errordetail:
                      - message: Program represented by Service-id is not enabled for serving Token Data. Please contact Mastercard administrator to enable this feature.
                        reason: INVALID_VALUE
                        source: serviceId
                        sourceType: BODY
                    message: Access is denied to the requested resource. Client does not have sufficient permission or has been locked.
                    reason: PERMISSION_DENIED
                    status: 403
                PERMISSION_DENIED-NOT_FOUND:
                  description: |
                    no active key returned from tcs / in_use key invalid / fringerPrints not matched
                  value:
                    errordetail:
                      - message: No active key found for this Client.
                        reason: NOT_FOUND
                        source: srcClientId
                        sourceType: BODY
                    message: Access is denied to the requested resource. Client does not have sufficient permission or has been locked.
                    reason: PERMISSION_DENIED
                    status: 403
              schema:
                $ref: '#/components/schemas/Error'
          description: |
            Forbidden.
        "404":
          content:
            application/json:
              example:
                errordetail:
                  - message: null
                    reason: null
                    source: Card
                    sourceType: BODY
                message: Unknown Error
                reason: NOT_FOUND
                status: 404
              schema:
                $ref: '#/components/schemas/Error'
          description: |
            Card not found by cards service.
        "500":
          $ref: '#/components/responses/InternalServerError'
        "503":
          $ref: '#/components/responses/ServiceUnavailable'
      tags:
        - Transactions
components:
  responses:
    InternalServerError:
      description: |
        Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            message: Internal server error. Something went wrong while processing the request.
            reason: INTERNAL
            status: 500
    ServiceUnavailable:
      description: |
        Service unavailable. Typically the server not able to serve the request temporarily.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            errordetail:
              - message: Issuer did not respond in time. Retry the request.
                reason: NO_RESPONSE_FROM_ISSUER
            message: Service unavailable. Typically the server not able to serve the request temporarily. Retry after some time
            reason: UNAVAILABLE
            status: 503
  parameters:
    ServiceId:
      in: query
      name: serviceId
      description: A unique identifier assigned by Mastercard for which tokens are created uniquely for the entity onboarded.
      example: "SECURE_COF_MERCHANT#START_COF-UPN2#01"
      required: true
      schema:
        $ref: '#/components/schemas/ServiceId'
    SrcClientId:
      in: query
      name: srcClientId
      required: true
      example: "6aff5cba-3aac-4fe0-b0cd-a52fc948d0c6"
      description: A unique identifier assigned by Mastercard during onboarding which signifies the responsible party Integrating to Mastercard Checkout Solutions (MCS).
      schema:
        $ref: '#/components/schemas/SrcClientId'
    SrcCorrelationId:
      in: query
      name: srcCorrelationId
      required: false
      example: "ff97ef9b-4b1e-40c2-b522-42197061dc5a"
      description: "A unique identifier that correlates a series of two or more requests to a single session of activity. Mastercard Checkout Solutions (MCS) will return a new srcCorrelationId in each response by default, but Integrators may choose to populate previously used srcCorrelationIds in subsequent requests to correlate their activity under a single ID. This can be done by providing a Mastercard generated srcCorrelationId in the request, or by generating your own ID in the same format. SrcCorrelationId is used for tracking and troubleshooting purposes within Mastercard's ecosystem."
      schema:
        type: string
        pattern: "^([0-9a-f]{8}.)?[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$"
    SrcDpaId:
      in: query
      name: srcDpaId
      description: |
        The registered identifier for the Digital Payment Application (DPA) accessing the service.
        __Conditional__: Required if you are a Merchant directly integrating with Mastercard Checkout Solutions (MCS) APIs, a Payment Service Provider (PSP) integrating On-Behalf-Of (OBO) a Merchant, a Payment Facilitator, or if you are enrolling in the Secure Card on File (SCOF) QR program.
      example: "5e0d4b84-189d-4c86-822d-590602f62062_dpa1"
      schema:
        type: string
    SrciTransactionId:
      in: query
      name: srciTransactionId
      description: Transaction reference generated by the checkout initiator
      required: false
      example: "18df482e-31f2-4289-b0d7-4d768b72d3ed"
      schema:
        type: string
        maxLength: 255
  securitySchemes:
    appInstanceIdCookieAuth:
      type: apiKey
      in: cookie
      name: appInstanceIdCookieAuth
    jwtBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Bearer . Carries one of the supported authorization types, structured as a JWT, generated and signed by an Identity Provider. Used to identify the associated customer account profile, and attests that the requester is authorised to access this data."
    xOpenApiClientId:
      type: apiKey
      in: header
      name: x-openapi-clientid
      description: Open API key to make the server to server call to Mastercard SRC System.
  schemas:
...
    DeleteCardResponse:
      type: object
      required:
        - srcCorrelationId
      properties:
        srcCorrelationId:
          $ref: '#/components/schemas/SrcCorrelationId'
...
    Error:
      type: object
      required:
        - status
        - reason
        - message
      properties:
        errordetail:
          description: List of Errors
          items:
            $ref: '#/components/schemas/ErrorDetail'
          type: array
        message:
          description: Error message details
          example: "correlation-id is required"
          type: string
        reason:
          $ref: '#/components/schemas/ErrorReason'
        status:
          description: HTTP response status codes.
          example: 400
          format: int32
          type: integer
    ErrorDetail:
      type: object
      required:
        - reason
      properties:
        reason:
          $ref: '#/components/schemas/ErrorDetailReason'
        message:
          description: Message displayed as result of error.
          example: "correlation-id is required"
          type: string
        source:
          description: Parameter that has the error.
          example: "correlation-id"
          type: string
        sourceType:
          $ref: '#/components/schemas/ErrorSourceType'
...
    ErrorSourceType:
      type: string
      description: Source of error.
      enum:
        - BODY
        - HEADER
        - QUERY
        - PATH
    GetCardResponse:
      type: object
      required:
        - maskedCard
      properties:
        srcCorrelationId:
          $ref: '#/components/schemas/SrcCorrelationId'
        maskedCard:
          $ref: '#/components/schemas/MaskedCard'
        encryptedPaymentData:
          description: |-
            This object is subject to manual set-up.
            Object for encrypted payment data, including paymentToken and relevant token information. Integrators must work with their account managers to be able to receive a token in the get card response.
          example: "eyJ0eXBlIjoiSldUIiwiYWxnIjoiUlMyNTYiLCJraWQiOiI4NTIxNDY4OCJ9.eyJzdWIiOiI1OTY3OGJmNS00ODcxLTQ5M2MtOWUwYi0"
          type: string
        keyFingerprintId:
          description: "Mastercard will use the encryption key associated with the supplied KID to encrypt the payload. If not supplied, the key which is marked as default will be used."
          example: "nwzNuN9upxolVsr6q0I/phcnfA/ZlaJ2gmAJiogMCwM"
          type: string
      description: The GetCardResponse provides card information associated with a srcDigitalCardId.
    GetRecentTransactionsResponse:
      type: object
      properties:
        transactions:
          items:
            $ref: '#/components/schemas/Transaction'
          type: array
          description: "All the transactions (last 10 transactions within the previous 30 days) for the requested Token (CardID), filtered by the ‘transactionsFromTimestamp’ if one was provided in the request."
...
    ServiceId:
      type: string
      maxLength: 255
      description: A unique identifier assigned by Mastercard for which tokens are created uniquely for the entity onboarded.
      example: "SECURE_COF_MERCHANT#START_COF-UPN2#01"
    SrcClientId:
      type: string
      format: UUID
      pattern: "[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}"
      description: A unique identifier assigned by Mastercard during onboarding which signifies the responsible party Integrating to Mastercard Checkout Solutions (MCS).
      example: "5e0d4b84-189d-4c86-822d-590602f62062"
    SrcCorrelationId:
      type: string
      pattern: "^([0-9a-f]{8}.)?[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$"
      description: "A unique identifier that correlates a series of two or more requests to a single session of activity. Mastercard Checkout Solutions (MCS) will return a new srcCorrelationId in each response by default, but Integrators may choose to populate previously used srcCorrelationIds in subsequent requests to correlate their activity under a single ID. This can be done by providing a Mastercard generated srcCorrelationId in the request, or by generating your own ID in the same format. SrcCorrelationId is used for tracking and troubleshooting purposes within Mastercard's ecosystem."
      example: "AeCC42A6-6Afd-cF73-c100-9a09CaCBceFd"
...